+++ tools/gen_mil_schema.py
@@ -65,6 +65,8 @@ MESSAGES = (
     "DictionaryFeatureType",
     "StringFeatureType",
     "EnumeratedShapes",
+    "ShapeRange",
+    "SizeRange",
     "Shape",
 )